Home

Baseball.jl

Baseball.absolute_pressure

Baseball.air_density

Baseball.vapor_pressure_over_water

Baseball.absolute_pressure — Method.

absolute_pressure(p0, h)

Calculate the absolute pressure [Pa] at altitude h [m] for a given sea level static pressure [Pa].

source

Baseball.air_density — Method.

air_density(P, T, ϕ)

Calculate density of humid air [kg/m^3] given absolute pressure P [Pa], temperature T [K] and relative humidity ϕ [1].

source

Baseball.vapor_pressure_over_water — Method.

vapor_pressure_over_water(T)

Compute the saturation vapor pressure over liquid water in Pa given temperature T in °C.

References

Arden L Buck (1981) - New Equations for Computing Vapor Pressure and Enhancement Factor

Arden L Buck (1996)

source